Allow changing status interval per task
Allow to change the status interval on the task detail page, i.e. after a task is already created. Previously one could only change this default value upon task creation.
Note that once a client is running a task, it does not yet automatically adapt the new statusTimer value. A pop-up message reminding the user of this property is shown when the status interval is changed.