* User-api / Abort chunks group
Allow to abort all chunks — with hashcat status 'running' or 'init'. (See src/inc/defines/hashcat.php) — belonging to all agents in a specified access group. This forces every agent in the group to drop his current chunk and request a new one, which in turn causes the server to automatically check whether there exists a task with a higher priority.

```
 The client requests chunks to be aborted for agents belonging to a specific access group.

    {
    "section": "group",
    "request": "abortChunksGroup",
    "groupId": "1",
    "accessKey": "<myKey>"
    }

    The server should reply following on success:

    {
        "section": "group",
        "request": "abortChunksGroup",
        "response": "OK"
    }
```