* User-api / Set task to top priority
Allow to set priority of tasks and supertasks to top priority within the set of pretasks. When setting the a task to be top priority, add one hundred to the previous highest priority and use this value.

```
 The clients sets a task to top priority.

 {
     "section": "task",
     "request": "setTaskTopPriority",
     "taskId": "1",
     "accessKey": <myKey>
 }

 The server should reply following on success:

 {
     "section": "task",
     "request": "setTaskTopPriority",
     "response": "OK"
 }
```

```
The clients sets a supertask to top priority.

{
    "section": "task",
    "request": "setSupertaskTopPriority",
    "supertaskId": "1",
    "accessKey": <myKey>
}

The server should reply following on success:

{
    "section": "task",
    "request": "setSupertaskTopPriority",
    "response": "OK"
}
```