Add uniform resource URI to objects
When creating objects and/or using expansion it is usefull to have a
uniform resource URI, this will avoid building manual queries. This
is exposed in the the special field '_self'.

While here also include the special field '_id' which is the shorthand
id required when linking to the object.